refactor: create component 'audioButton'
