+++ src/pages/textBook/textBook.scss
@@ -237,5 +237,4 @@ nav {
 .word-card_audio-button-image {
   width: 3rem;
 }
-
 //min-width: 1200, 992, 768, 576, 320
